file moved to specific folderThose that belong to raceMap are moved to raceMap folder those to mainMap to MainMap
