+++ src/components/auth/Signup.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useState } from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../context/AuthContext";
 import { useNavigate, Link } from "react-router-dom";
 
 function Signup() {
