+++ src/components/raceMap/RaceMap.jsx
@@ -1,7 +1,7 @@
 import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
 import { useState, useEffect } from 'react';
 import { ref, onValue } from 'firebase/database';
-import { database } from '../firebase-config';
+import { database } from '../../firebase-config';
 
 const RaceMap = () => {
   const [positions, setPositions] = useState({});
